+++ src/domain/index.ts
@@ -14,3 +14,4 @@ export * from "./touchpoints";
 export * from "./intent-progress";
 export * from "./brand-asset-kinds";
 export * from "./connector-result";
+export * from "./overton-radar-signal";